media: capture: Do not spend time in Camera::requestComplete handler
When a request is completed, the Camera::requestComplete() callback runs in the
CameraManager thread. Handling of the completion must be as fast as possible,
and must not block. The bulk of the work must be thus be moved to a separate
thread. Post a task to the camera_task_runner_ to handle the completed request
in a separate CameraCaptureDelegate::ProcessRequest() function, instead of
doing so synchronously in CameraCaptureDelegate::RequestComplete().

This creates two race conditions at stop time that must be considered.

CameraCaptureDelegate::StopStream() is run in the camera_task_runner_. As the
task runner is a sequential tast runner, it guarantees that tasks will be run
sequentially. The order in which tasks are executed is the order in which they
are posted. Tasks are posted to the runner from two different threads, in
VideoCaptureDeviceCamera::StopAndDeAllocate() and in
CameraCaptureDelegate::RequestComplete(). We thus can't fully control in which
order tasks are posted when stopping.

In VideoCaptureDeviceCamera::StopAndDeAllocate(), two tasks are posted
sequentially, the first one to stop the camera by running
CameraCaptureDelegate::StopAndDeAllocate(), and the second one to delete the
CameraCaptureDelegate instance (by a call to the task runner's DeleteSoon
function, which, at least conceptually speaking, posts a task to the runner).
There's thus a guarantee that CameraCaptureDelegate::StopAndDeAllocate() will
complete before the delegate is destroyed.

As requests complete asynchronously relative to the stop operation, we need to
consider when the task posted by CameraCaptureDelegate::RequestComplete() is
inserted in the queue, relatively to the two tasks from
VideoCaptureDeviceCamera::StopAndDeAllocate().

- If the request completion task is schedule before
  CameraCaptureDelegate::StopAndDeAllocate(), everything is fine, the request
  is handled before the camera is stopped.

- If the request completion task is scheduled after
  CameraCaptureDelegate::StopAndDeAllocate() and before DeleteSoon, the request
  will be deleted by CameraCaptureDelegate::StopAndDeAllocate() before the
  completion task runs. The completion task will thus access an invalid
  request, we have a race condition. The CameraCaptureDelegate instance,
  however, is still valid at that point as DeleteSoon hasn't run.

  We handle this by checking at the beginning of the request completion task
  (CameraCaptureDelegate::ProcessRequest()) if the capture is running. If it
  isn't, it means that CameraCaptureDelegate::StopAndDeAllocate() has been
  called and has completed, and the request pointer isn't valid anymore. We
  thus return immediately in that case, as no cleanup is required.

- If the request completion task is scheduled after DeleteSoon, not only will
  the request be invalid, but the CameraCaptureDelegate instance will also have
  been freed.

  We handle this by passing a weak pointer to BindOnce when posting the request
  completion task. The pointer will be set to null when the
  CameraCaptureDelegate is deleted, and the task runner will then ignore the
  task. We're not leaking any resource by doing so, as
  CameraCaptureDelegate::StopAndDeAllocate() has already freed the request and
  all other resources on the libcamera side.

Signed-off-by: Jean-Michel Hautbois <jeanmichel.hautbois@ideasonboard.com>
Signed-off-by: Laurent Pinchart <laurent.pinchart@ideasonboard.com>